requirements.txt
gdown
ultralytics
torch
torchvision
opencv-python
Pillow